split security extensions implementation into multiple files
